chore: pin `pnpm` for corepack (#1831)
