+++ .github/workflows/release.yml
@@ -16,15 +16,13 @@ jobs:
         uses: actions/checkout@v3
 
       - uses: pnpm/action-setup@v3
-        with:
-          # https://github.com/pnpm/pnpm/issues/8953
-          version: 9.15.3
 
       - name: Use Node.js from nvmrc
         uses: actions/setup-node@v4
         with:
           node-version-file: ".nvmrc"
           registry-url: "https://registry.npmjs.org"
+          cache: "pnpm"
 
       - name: Install Dependencies
         run: pnpm i --frozen-lockfile
