+++ package.json
@@ -40,6 +40,8 @@
   },
   "dependencies": {},
   "engines": {
-    "pnpm": "^9.0.0"
-  }
+    "pnpm": "^9.15.0",
+    "node": ">=22"
+  },
+  "packageManager": "pnpm@9.15.0"
 }
