+++ packages/tests/package.json
@@ -23,9 +23,9 @@
     "@vitest/ui": "3.0.5",
     "jsdom": "^25.0.1",
     "lodash": "^4.17.21",
-    "solid-js": "^1.9.4",
+    "solid-js": "^1.9.5",
     "vinxi": "^0.5.3",
-    "vite-plugin-solid": "^2.11.1",
+    "vite-plugin-solid": "^2.11.6",
     "vitest": "3.0.5"
   },
   "engines": {
